Sockets, threading, multiple servers, config file
Additions:
------------
- I added support for multithreading
- I added support for multiple servers
- I added code to load and use the config file
- Replaced the irclib with sockets (Sorry frank, but it would have taken
longer to understand how that library worked also, it would make the
code more OOP-oriented)

Things to look out for:
-------------------------
- PINGing might not work, will look to fix that in my next revision if
it needs to be
- No easy way to close out of the bot, not even a keyboard exception
(Ctrl + Z) will terminate the bot. Threads are running independent of
the main thread, so I'll soon implement something to solve that.
- No comments :P I'll add some soon

Things to add:
----------------

- Support for plugins and commands, we need them.
- SASL and other forms of authentication
- Official, "tried and true" way of daemonizing bot